requirements.txt
fastapi
uvicorn
fastapi-mail
jinja2
celery
redis
python-dotenv
aiosmtplib
